pyproject.toml
[build-system]
requires = ["hatchling", "versioningit"]
build-backend = "hatchling.build"

[project]
name = "tagkit"
description = "A CLI & Python API for editing EXIF data in photos."
readme = "README.md"
authors = [
    { name = "Robert Forest", email = "rvforest@gmail.com" }
]
requires-python = ">=3.9"
dynamic = ["version"]
dependencies = [
    "piexif>=1.1.3",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "typer>=0.15.2",
]

[dependency-groups]
    dev = [
    "nox>=2025.2.9",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.6",
]

[project.scripts]
tagkit = "tagkit.cli.main:app"
    
[tool.ruff]
per-file-ignores = { "__init__.py" = ["F401"] }

[tool.hatch.version]
source = "versioningit"

[tool.versioningit.vcs]
method = "git" 
match = ["v*"]
default-tag = "0.1.0"

[tool.versioningit.next-version]
method = "smallest"

[tool.versioningit.format]
distance = "{next_version}.dev{distance}+{vcs}{rev}"
dirty = "{base_version}+d{build_date:%Y%m%d}"
distance-dirty = "{next_version}.dev{distance}+{vcs}{rev}.d{build_date:%Y%m%d}"

[tool.versioningit.template-fields]
pep440 = true

[tool.coverage.run]
omit = [
    "noxfile.py",
    "tests/*"
]
